chore(core): update latlng data type

+++ core/src/main/kotlin/dev/ikti/core/data/local/entity/LocalUserEntity.kt
@@ -16,7 +16,7 @@ data class LocalUserEntity(
     @ColumnInfo(name = "telepon") val telepon: String,
     @ColumnInfo(name = "foto") val foto: String,
     @ColumnInfo(name = "alamat") val alamat: String,
-    @ColumnInfo(name = "alamat_lat") val alamatLat: Float,
-    @ColumnInfo(name = "alamat_lon") val alamatLon: Float,
+    @ColumnInfo(name = "alamat_lat") val alamatLat: Double,
+    @ColumnInfo(name = "alamat_lon") val alamatLon: Double,
     @ColumnInfo(name = "foto_pegawai") val fotoPegawai: String
 )
